Remove ActivityListViewModel
